+++ src/collections/Media.ts
@@ -5,6 +5,7 @@ const Media: CollectionConfig = {
 	fields: [
 		{
 			name: 'alt',
+			label: 'Alt Text',
 			type: 'text',
 		},
 	],