+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2023-04-22
+
+### Added
+
+- `GolfCourses` Collection
+- `Properties` Collection
+
+### Fixed
+
+- Missing `publishedDate,` `slug,` `author,` and `status` on multiple collections.
+
+### Changed
+
+- Navigation & Subnavigation Updates
+- 'versions' Added to All Collections
+
+### Removed
+
 ## [1.0.0] - 2023-04-12
 
 ### Added